[#18]style: 서비스단 기능별 분리
