+++ src/main/java/com/sudurukBackBack/Modu_Lecture/domain/storage/exception/LocalFileUploadException.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 /**
  * 파일 저장 중 발생하는 예외를 나타냅니다.
  */
-public class FileStorageException extends BasicException {
+public class LocalFileUploadException extends BasicException {
 
     @Override
     public int statusCode() {
@@ -15,6 +15,6 @@ public class FileStorageException extends BasicException {
 
     @Override
     public String errorMessage() {
-        return "로컬 비디오 파일 업로드 과정에서 문제가 발생하였습니다.";
+        return "로컬 파일 업로드 과정에서 문제가 발생하였습니다.";
     }
 }
